@@ -18,10 +18,12 @@ skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
 known_third_party = ["azure", "cognite", "prometheus_client"]
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<3.11"
+python = ">=3.9,<3.11"
 cognite-sdk = "^6.26.0"
 cognite-extractor-utils = "^5.4.1"
-azure-eventhub = "^5.11.4"
+azure-identity = "^1.15.0"
+deltalake = "^0.15.3"
+pandas = "^2.2.1"
 
 [tool.poetry.dev-dependencies]
 pyinstaller = "^4.7"
